Improve false handling. Making batch loading of settings configurable
Don't use a key value array for false handling - array keys aren't that
sensitive.

It's likely that batch loading is faster than looking for each setting
one by one. However, make it possible to disable batch loading without
changing code in case it's not the case for a project - or in general
(to be proven/demonstrated)